Rework handling of bundle preferences. Remove the includeAllSettings parameter to preferenceItems. Instead provide a version that takes an array of preference keys and returns a dictionary with scopes as keys with values being dictionary of attribute-value pairs.
The old preferenceItems method is renamed preferenceItem (singular). It
returns a dictionary with scopes as keys and a preference value as values.

This fixes cases where symbolTransformation preferences are not grouped
together with a showInSymbolList preference. In either case, the
includeAllSettings code was a misunderstanding.

This commit implements support for additional theme attributes in bundles.
The Diff bundle adds background, foreground and fontStyle preferences,
which are now merged in with the other theme attributes.

Preference settings are now normalized at load time, ie color settings are
transformed to a proper NSColor object, with keys appropriate for an
NSAttributedString. This now works for all bundle preferences too.

Reenable caching of theme attributes per scope. I don't know why it was
disabled, I should really write better commit messages!