Change how web views are reused in bundle commands.
Associate a web view (command output controller) with a document and a
bundle command. Reuse associated web views in repeated invocations of
the bundle command on the same document.

This works well for preview commands that are logically associated with
a document. Another mechanism is probably needed for project-level
output that are invoked equally on multiple documents.

Reuses an existing web view that have been moved to another tab or
another window.