Add support for a custom status line.
Currently you have to manually create ViStatusComponents and add them to
the ViStatusView that is used to display the window status. There is a
convenience class, ViStatusLabel, that is used to display simple text.
Since the purpose of the status line is scriptability, many of the
custom components will be implemented in Nu and react to Vico events.

ViStatusComponents can be aligned left, right, or center. There is also
an automatic setting that simply leaves the component aligned in the way
that the previous component is aligned, unless it is centered. I'm still
figuring out the automatic setting and whether it's going to stay or
not. When setting the status components, you can include a @"%=" in the
array to make any subsequent components be forcibly right-aligned. This
allows you to not explicitly set alignment properties on the components
if you want to avoid it.

Also coming is statusline string support, which will parse a statusline
string and turn it into a series of ViStatusComponents.