Only update control when value changes in ViStatusNotificationLabel.
Before, we would run a setStringValue and adjustSize on every
notification that we received. Now, when we get a notification, we check
the value before and the value after the notification, and we only run
setStringValue and adjustSize if they are different.

Amongst other things, this means ViStatusLineUpdatedNotification is only
posted if there is an actual change in the status line.