Invalidate sizes of affected components when adjusting.
When we adjust the size of a right-aligned component, we invalidate the
size of all components to our left that are right-aligned. When we
adjust the size of a left-aligned component, we invalidate all
components to the right that are left-aligned.

This commit also ensures that no component is left with automatic
alignment after they are added to the list initially. The initial
addition process assigns the correct alignment as it iterates through
the components.