Extend Nu support with a transformer block in ViStatusNotificationLabel.
ViStatusNotificationLabel updates a status label based on an Objective-C
notification. The transformer block allows for a Nu block to be run to
transform the notification into a String. It is only run if it is set on
a given label.

Additionally, an instantiation function is provided,
initWithNotification:transformerBlock:, that allows a
ViStatusNotificationLabel to be created with a transformer block right
off the bat.