Make default status bar be speicified in Nu.
app/status.nu now contains the definitions for the caret and editor mode
labels. The ViAppController has a statusSetupBlock property that takes a
Nu block that is called whenever a new window loads. The block receives
as its parameter the ViStatusView for the new window, and can then set
up the status view as it pleases.

The default status view has the caret and current mode. There is a
fallback in place where if the status block is nil, the caret and editor
mode labels are hardcoded in Objective-C. This fallback should only end
up being used if the user explicitly unsets the statusSetupBlock
property, since the Nu code bundled with Vico sets a default
statusSetupBlock.