Add notification for busy mode change.
When a ViDocument needs to handle a bundle command, it needs to change
the document state to busy. This should trigger a mode change for the
status bar and other areas of the system; however, the ViDocument can't
generate the ViModeChangedNotification because it doesn't have the
ViTextView to pass as the notification object. As such, we introduce an
intermediary notification, ViDocumentBusyChangedNotification, which is
used to notify any associated ViTextViews about the change, so that they
can in turn emit the appropriate ViModeChangedNotification.