Ignore mode changes triggered by other windows.
Before, all status views would reflect the mode of the current window's
ViTextView, because the notification broadcast to all status views and
they did not filter by window. Now we check that the associated text
view's window and the window of the status view the label is in are the
same before making any changes to the text.