Give the completion view vico keybindings.
To do that, we introduce a new class, ExCompletionView,
which is used for the NSTableView that displays the
completion results. It uses a ViMap for ex completion that is
similar to the ack window map (in fact, they both include
the same base map, tableNavigationMap) with some
differences for the handling required by how completion
works in the face of the enter, tab, and esc keys.