Always report affected lines/number of replacements in :s.
Before, :s only reported the lines and replacements when
provided with the /n flag, which indicated a dry run instead
of a true replacement. It now reports them even when the
replacement is actually run.

Currently there are some issues updating the status line
when the :s is not provided with the /n flag that indicates a
dry run rather than a proper replacement.