Don't hide messages set in this run loop cycle.
Before, the new ViStatusView hid any message that was
being displayed whenever a status component updated.
The issue with this was that sometimes in one go of the run
loop you would both have a status component change (e.g.,
a changed caret) and a request to display a message. In
these cases the component change would typically win,
making it impossible to see the message.

We now temporarily disable automatic message hiding for
the remainder of this run loop cycle when the message is
set, so that other status component updates won't hide
the message.

We also remove some code that was marked TODO from
the original status view refactoring that was related to fixing
this issue in the old code.