Pass keys from completion controller to text view as well.
We pass in the text view's key manager when initializing
the completion controller, and it goes on to pass keystrokes
back to the text view. This way, we're updating the contents
of the text view as the user types, in addition to filtering the
completion view itself.

This replaces the old
completionController:appendedStringWithoutCompleting:
hack, which was totally broken anyway.