Trigger completion window based on autocomplete preference.
We also only trigger it once at a time. So if the completion
window is already up, we don't try to bring it back up, which
will reset its state. Instead, we track if we've currently got
a window up (which is easy since it runs modally) and don't
try again until it's gone.