Make space accept only if not autocompleting.
We add a selector, accept_if_not_autocompleting:, to
the completion controller. When invoked via autocomplete,
the completion controller gets a C option, which we check
for and then set on an instance variable. That variable is
then checked when invoking accept_if_not_autocompleting:
to see if the selected entry should be accepted or not.

The effect of this is that invoking autocomplete via C-N will
accept on space, since you explicitly asked for the
completion popup, while if it pops up via autocomplete it will
only accept on enter/tab. The additional effect is that this is
customizable behavior -- the relevant keys can simply be
remapped to accept: or accept_if_not_autocompleting:
as desired.