Pass delegate to chooseFrom: in ViCompletionController.
Before it was expected that the delegate would be set
before, but then ViCompletionController runs a reset when
it starts. This way we ensure that the completion controller
has the proper delegate while it's modally presented, and all
exit points reset the controller and clear out the delegate
by consequence, making this a safer flow.