Add a default catchall action to ViMap for non-character keys.
ViMap's default action only passes on a particular range of
keystrokes. For complete-as-you-type, we need to be able
to receive all keys so we can pass them on to the text view
key manager. Without the catchall default action, we end up
missing things like alt-backspace, and never take action.