Close autocomplete popup on >1 character changes.
When the user makes a change that is >1 character, like a
word deletion, we force-close the autocomplete popup. In
the completion controller, we ignore invocations to cancel:
if there is no delegate, since that should mean the
completion window isn't up.

Lasty, we don't process smartindent if the completion popup
is cancelled/aborted. This prevents an issue with out of
range strangeness if the user cancels a completion at the
end of the line by doing a delete word.