Properly update final_location after completion popup closes.
This ensures anyone else who relies on it in the rest of the
input handling chain (which is blocked by the completion
popup's modality) will have the right value. In this case, we
can just set final_location to the caret, since the completion
popup method takes care of updating the caret appropriately
when/if the text insertion step wraps up.