Remember key manager for autocomplete cancel.
Before, when the user cancelled the completion popup
during autocomplete, we tried to pass on the cancellation
key combo to the existing key manager. Unfortunately we
did this after the completion controller had been reset, which
meant there was no key manager to send it to. We now
save the key manager before resetting and then pass it the
keys after we run the reset. This is so that the key manager
receives the keys after the completion controller is fully
dismissed.