When cancelling in accept_if_not_autocompleting:, don't pass keys on.
We were passing the keys that cancelled the popup to the existing key
manager in situations where we were running
accept_if_not_autocompleting: and the completion popup *was*
autocompleting. This typically meant we'd get two spaces instead of just
one when we'd hit space, which should cancel the popup but not accept
the top completion when autocompleting.