Insert completions by using keyManager handleKeys:.
Completions are now inserted by virtually playing the keys
that the completion would add to the current text as actual
keys typed by the user. We set virtualInput to YES while we
do this. This allows the keys to be recorded for the dot
command to replay, which fixes dot commands with
autocomplete.