Update caret before triggering autocompletion.
Because the completion dialog is modal, processing for that
particular keystroke stops until the completion popup goes
away in the situation where we show the completion popup.
This means we never get a chance to show that the caret
moved to after the currently inserted character. We now
update the caret explicitly before showing the completion
popup so that it updates correctly in the UI. This fixes a
couple of other bugs with snippets, as well.