Add support for omni completion in language bundles.
Omni completion is specified in the form of a filename that
is searched for in the Completions/ subdirectory of the
bundle. It is then checked for functions
<bundle.name>.completionRange and
<bundle.name>.completions. The first one is passed the
location that the completion was requested at in the
document and returns the range of the token that should be
completed. The second will probably take the location and
the token to be completed and return a list of completions.