Track line/column per document instead of caret/scroll point.
Trying to set the scroll point too early in the document loading process
was causing exceptions and general nastiness. This is already handled
when you try to set the text view to a particular line/column, so we
lean on that instead and track the user's current location based on that
information.