Remove a possibly-not-useful guard in tab creation.
Originally, we were only creating a tab if the first document
we looked at at the top level was not also the only
document. It's unclear why that would be a necessary
condition, so that condition goes away until we can prove
that it's necessary (at which point we'll indicate why inline).