Select associated view when creating a tab for a document.
We weren't doing this before, which resulted in some
strange interactions when creating a tab from the project
loading logic. We would typically create an extra horizontal
split instead of loading correctly.

Once again, there's no clear reason why we wouldn't be
selecting the document view after creating the tab, and
barring an explanation, we're going to go ahead and do it.
Creating new tabs normally works correctly, as does
creating new tabs when the default for opening files is
doing it in a new tab.