Properly check return value of keyManager:shouldParseKey:inScope:.
That method was returning an NSNumber, but we were
checking it as if it were returning a BOOL (using ! <value>).
The result was that, since we always returned an
NSNumber instance, we never detected when a key was
being parsed.

We now properly unpack the NSNumber and check the
boolValue inside for NOness.