Switch project saving to use PSMTabBarControl information.
Before, we were using the NSTabView in the window
controller to save information on what tabs were open.
However, the information on what order the tabs are listed
in the tab bar is inaccurate in the NSTabView; we have to
use the PSMTabBarControl that is being used to display the
actual tabs to get that information accurately.

This fixes an issue where drag-and-drop reordering of tabs
after they were opened would not be reflected after closing
and then reopening a project.