Set the theme opacity on window initialization.
We were only updating the window opactiy based on the
theme when the theme changed; however, this meant that
when the application started it would always be opaque.

We now move the logic to set the window opactiy based on
the theme's background opacity into a dedicated method
that we call both from the theme change KVO handler and
when the window initially opens.