Temporarily make window key on opacity change if needed.
Before, changing the theme in the preferences to a theme
with a semi-transparent background would result in the
semi-transparency painting over a gray background instead
of the actual stuff behind the window until the window
becomes key.

To avoid this, when setting the window opacity, if the window
controller's window isn't key, we make it key temporarily and
then restore the original key window's key status.