Hack fix so complex capture highlighting doesn't kill Vico.
Right now a capture highlight that contains more than just
a name for the capture group makes Vico enter an
unrecoverable state. We guard for the `name` key being
present in the capture object before trying to use it.

This is a stopgap fix until we implement proper support for
more complex capture highlight definitions.