When looking for a handler for a command, fall back on completion view.
More broadly, this applies to the target for any selector. If
the completion view is showing and the text view can't
handle the given selector, we try to see if the completion
view can. This lets the completion controller deal with
selectors that the text view doesn't need to worry about.