Map movement keys to unique methods in ViCompletionController.
We don't want to be triggering the standard up/down
movement in the text view when the completion window is
up; instead, we want to trigger movement through the
completions themselves. Because of how we have the key
handling wired up now, to do this, we just map the key
combos to unique methods in ViCompletionController, rather
than mapping them to methods that have the same name in
ViCompletionController and ViTextView, and everything
takes care of itself.