Add completion window as child after updating its bounds properly.
addChildWindow:ordered: seems to show the window briefly
even if it is hidden when the method is invoked. Thus, we
rewire things in the completion controller to allow us to set
up the bounds before adding the window as a child of the
invoker, so that we don't get a brief flash of the window in a
bad position.