Turning line numbers off with word wrap no longer causes documents to be cut off.
Ultimately this was a pretty weird issue, related to the fact
that when we turn word wrap on we match the text view's
NSTextContainer's width to the clip view width, and then ask
the container's width to track the text view's.

When line numbers are on, we call setRulersVisible:YES on
the scroll view. This does some layout calculations and sets
up the text view's clip view properly before we deal with
word wrap. However, when line numbers are *off*, the
clip view dimensions aren't set up properly until *after* the
word wrap is handled.

To fix this, we hackishly wait until the next runloop cycle to
handle wrapping *specifically* in the case where word wrap
is being turned on while the text view is being initialized. If
word wrap is being turned off, or if this is an update after the
text view initialization (say, because the user changed a
preference), things work the usual way, so we don't delay
the invocation.