Add createFoldForRange: to ViDocument.
This registers a fold for a given range of content. Currently
it only deals with one level, setting a flag when there is a
fold. More sophistication to come.

Note that folds are linewise, so the range is resolved to a
set of line numbers and folding set on those lines.