Fix an issue where leaving insert mode with completion up wouldn't set the parser right.
Because by the time normal_mode: runs, the parser has
already been reset to normal mode, but the `mode` instance
variable hasn’t, nuking the completion controller at this
stage was causing the parser to be set back into insert
mode (because the completion controller cancel handler
in ViTextView restores the insert map on close so that the
completion map is no longer in effect).

To deal with this, we now cancel the completion controller
after the text view’s `mode` variable has been set.