Let addChildWindow:ordered: show the completion window.
Before we were explicitly doing an orderFront: first. This
meant that if we had been showing the completion window
in a different space before we tried to bring it up in this
window, the user would get pulled to the previous space and
then pulled back.

That’s obviously not cool ;)