Add exCommandCompletionMap.
Before, we were using the regular completion map in the
field editor; however, we need a different map for completion
in the ex command field, since the base keystrokes are
different.

Right now this is identical to the standard completion map
except that it includes the ex command map instead of the
insert map. A couple of attempts to create a base
completion map with the completion keystrokes and then
include that in addition to either the ex map or the insertion
map depending on which completion map we were in didn’t
seem to go so well, so we’ll move on for now.