Only exclude closed child folds from being unmarked as folded.
This when closing a parent fold. Before, we would
indiscrimately skip child folds when removing the folded
attribute, even if the child fold had never been closed. This
led to some interesting-looking results, since the first
character of the child fold had never been marked with the
folded attachment, so you just ended up with the first
character of the child fold and none of the other characters
therein.