Semi-functioning updating of folds when content changes.
Something is still getting futzed up on the beginning location
of the fold if you edit within it, even though it’s expanding
properly. There’s also some odd behavior at the edges, e.g.
O on the first line of a fold extends the range while o above
that line does not, and o on the last line of the fold reduces
the range’s length by one line. More to come.