Fix check for overlapping ranges when creating a fold.
Because enumerateAttribute:inRange:options:usingBlock:
clips the ranges it reports to the requested range, our check
to see if the overlapping fold range was greater than the
requested fold range was going to always be false. We now
check for equality, which indicates that the overlapping fold
is >= the length of the requested fold.