Tweak open_line_below to operate "properly".
open_line_below was previously inserting a newline before
the terminating newline of the current line. We now properly
insert the newline AFTER the terminating newline of the
current line. This is the "correct" behavior, and lets us look
at the character where the insertion is occurring to decide
what fold to use as with other situations.