Delay updating of ruler rule thickness by a run loop.
This was in the code that we displaced when we
implemented the fold margin view, but got lost in the
transition. Without the runloop delay, the layout manager will
throw some seriously nasty exceptions trying to access
no-longer-existing characters while re-laying out the text
view in response to the thickness change.