Only mark the start of a closed fold when opening its immediate parent.
We were marking the start of a closed fold with the folded
attachment if we had opened any parent, which means if
we had folds closed three levels deep and we opened the
top level, we’d see a marker both for the second and third
level closed folds. This fixes that issue.