closestCommonParent fold now checks if its arguments are parents of each other.
We were running into an issue where if the first fold was a
parent of the second, and they both shared another parent,
you would get the other shared parent, the first fold’s parent,
as the result.

Now, if the first fold is parent of the second, it is returned,
and if the second is parent of the first, it is returned, and if
not then we search for the nearest common parent as we
did before.