Only try to find a parent fold to close if there is room to do so.
Before, when closing a fold at a given location, we would
push through parents even if the current fold’s depth was
equal to the minimum close depth, which meant we could
try to close folds above the minimum depth. We no longer
do that.

This only happened when folds had the same start location
as their parents and we were closing one of the child folds.