Only mark the first character of a folded range with attachments.
Before we were checking if the current fold was the same as
the first fold, but that is also true if we go through a child fold
and then back to the first fold in a later range. In those
cases we don’t want to mark the first character in the current
range with the attachment attribute.

End result, we now just check if we’re closing this fold and
the start of the current range is the start of the fold area.
In that case, we mark it with the attachment attribute.