Compute and constrain reference point with respect to scroll view.
We now compute the reference point as a point in the scroll
view containing the text view. Additionally, we constrain it to
the scroll view’s bounds with some padding, so that if the caret
is off the visible scroll area, we keep the value we’re working
with within the scroll rectangle. That means we’ll never be able
to switch windows outside of the visible area of the current
view.