While searching for a view to move up/down/left/right to, adjust for splitters.
Before, if going in the indicated direction from the caret
resulted in coordinates that were on a split view splitter, we
would simply fall back on the old behavior of going to the first
entry in the split in the given direction. We now adjust the
coordinates either rightwards or downwards (depending on
whether we’re looking for the next text view vertically or
horizontally) so as to move off of the splitter.