When searching for a view to move up/down/left/right to, recognize ruler views.
Before, hitting a ruler view would mean continuing to adjust in
the given direction, which when moving up or down would
mean continuing to look up or down for a valid text view. In
those cases, we wouldn’t end up finding one, or we’d end up
passing through the view immediately on top/bottom of the
current one to one further up whose ruler wasn’t in the path of
the caret point.

Now, when we see a ruler view or one of its children, we look
up the associated text view and select that.