Add syntaxHintFor:withPrefix: to ExMap.
syntaxHintFor: takes all of the possible aliases for
ExMapping and assumes the required prefix is at the
longest the shortest available alias. However, this doesn't
work when not all mappings are prefixes of each other, for
example with tabedit and tabnew. In these cases, we
introduce the withPrefix: parameter, which can be used to
provide a prefix that would choose between the available
aliases for the given ExMapping.

For example, calling syntaxHintFor: with the above tabedit
mapping would yield tab[edit] with no prefix or with any
prefix of tabedit. However, if you pass tabn as the prefix,
the command hint will be tab[new] instead.