Rework Oniguruma for Swift compatibility.
Oniguruma by default is designed to traipse all over regex_t and other
POSIXy APIs. Swift doesn’t like this one little bit, and we don’t use
that at all in Vico, so we drop the compatibility headers responsible
for traipsing as well as the corresponding ifndef that sets up the
initial regex_t alias.