Provide typed access to untyped Dictionary values.
Right now, some Cocoa dictionaries are of AnyObject types. For example,
NSNotification’s userInfo is a [String: AnyObject]. We extend
Dictionary to provide safe typed get operations on these, which, when
you request a T, will return a T?, nil if the key is either missing or
has the wrong type.

We also provide a typedGet for Int that will automatically unwrap an
NSNumber.