Start reworking ruler view and line number view.
We rework ViRulerView in as a parent view with subviews. These are
managed by ViRulerHelpers, of which there is currently only a
ViLineNumberView. These in turn provide access to helperViews which are
in charge of rendering the relevant contents.

Right now ViLineNumberView is both a view and a helper. Its rendering
is not yet perfect, and is somewhat slow.