Introduce ViRulerHelperConfig.
ViRulerHelperConfig represents ViRulerView’s helpers and their layout.
It’s also in charge of installing that layout, and of providing a
default set of helpers.

While implementing this, we switch all of ViRulerView to auto-layout,
and manage the line number view as a subview. In the process, we get
rid of a lot of now-unnecessary needsDisplay calls. We also move a lot
of responsibility for notifications and text attribute management
inside of the helper views and out of ViRulerView.

In general, everyone’s responsibilities are like a million times better!